+ Renamed project files. + Removed CPlayer type and Player files. + Implemented enemy AI. + Improved combat functions to support enemy AI to have the same actons available as the player character. + Moved inventory to CCharacter class so enemies have inventory too. + Implemented strength and focus potions. + Added size() function to query static array sizes. + Added SCombatBonus struct as member of CCharacter for holding modifiers for in-combat points. + Renamed SGameCounters to SCharacterScore and added it as member of CCharacter. + Made enemies drop items from the inventory on death. + Added TURN_OUTCOME, TURN_ACTION and ATTACK_TARGET enumerations for resolving the combat events. + Removed CurrentEnemy from CGame class because it's now created temporarily in the combat() function.

+++ KitsuRPG/Menu.h
@@ -27,7 +27,7 @@ int displayMenu(size_t optionCount, const std::string& title, const SMenuItem (&
 		
 		// Get user input
 		// Convert the input string to a menuitem index.
-		const unsigned int selectedIndex = getNumericInput()-1;	// 
+		const uint32_t selectedIndex = (uint32_t)(getNumericInput()-1);	// 
 
 		if(selectedIndex >= optionCount)	// We only accept from 0 to optionCount
 			printf("Invalid answer. Answer again...\n");
